+++ vcd/resource_vcd_snat.go
@@ -2,11 +2,11 @@ package vcd
 
 import (
 	"fmt"
+	"log"
+
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/hashicorp/terraform/helper/validation"
-	"github.com/vmware/go-vcloud-director/v2/govcd"
 	"github.com/vmware/go-vcloud-director/v2/types/v56"
-	"log"
 )
 
 func resourceVcdSNAT() *schema.Resource {
@@ -87,7 +87,7 @@ func resourceVcdSNATCreate(d *schema.ResourceData, meta interface{}) error {
 			return fmt.Errorf("error creating SNAT rule: %#v", err)
 		}
 	} else if networkName != "" && networkType == "ext" {
-		externalNetwork, err := govcd.GetExternalNetwork(vcdClient.VCDClient, networkName)
+		externalNetwork, err := vcdClient.GetExternalNetworkByName(networkName)
 		if err != nil {
 			return fmt.Errorf("unable to find external network: %s, err: %s", networkName, err)
 		}
@@ -152,8 +152,8 @@ func resourceVcdSNATRead(d *schema.ResourceData, meta interface{}) error {
 		if orgVdcNetwork != nil {
 			d.Set("network_type", "org")
 		} else {
-			externalNetwork, _ := govcd.GetExternalNetwork(vcdClient.VCDClient, natRule.GatewayNatRule.Interface.Name)
-			if externalNetwork != nil && externalNetwork != (&govcd.ExternalNetwork{}) {
+			externalNetwork, _ := vcdClient.GetExternalNetworkByNameOrId(natRule.GatewayNatRule.Interface.Name)
+			if externalNetwork != nil {
 				d.Set("network_type", "ext")
 			} else {
 				return fmt.Errorf("didn't find external network or org VCD network with name: %s", natRule.GatewayNatRule.Interface.Name)
@@ -245,8 +245,8 @@ func resourceVcdSNATUpdate(d *schema.ResourceData, meta interface{}) error {
 		natRule.GatewayNatRule.Interface.HREF = orgVdcNetwork.HREF
 
 	} else if d.Get("network_type").(string) == "ext" {
-		externalNetwork, _ := govcd.GetExternalNetwork(vcdClient.VCDClient, d.Get("network_name").(string))
-		if externalNetwork == nil || externalNetwork == (&govcd.ExternalNetwork{}) || err != nil {
+		externalNetwork, err := vcdClient.GetExternalNetworkByName(d.Get("network_name").(string))
+		if err != nil {
 			return fmt.Errorf("unable to find external network: %s, err: %s", networkName, err)
 		}
 		natRule.GatewayNatRule.Interface.Name = externalNetwork.ExternalNetwork.Name